Crate flowstdlib

flowstdlib is a library of flows and functions that can be used from flows.

The flow and function definition are used at compile time when compile flows that reference it.

At run-time, library entries can be of two types, indicated by their ImplementationLocator

Native - a native binary (containing) all functions is built and linked by a flow runner program (e.g. flowr) so that any function referenced by a flow is executed natively at native speeds. flowr offers the user control using this via the -n, --native command line option.

RelativePath - functions are compiled to WASM files and located within the library at runtime by the flow runner using this file path relative to the lib root. If either the library if not linked natively, or the -n, --native command line option is not used, when the function is referenced by a flow being run, it is loaded and executed in a WASM runtime.

Modules

control

Functions and flows to control the flow of data in a flow based on control inputs. functions for controlling data flow

data

Some generic processes that act on data. functions for generic operations on data

errors

provides Error that other modules in this crate will use errors::*;

fmt

Functions for the formatting of values and conversion from one type to another. functions for formatting data

manifest

Use manifest::get_manifest to get the natively/statically linked LibraryManifest for this library to get access to everything error_chain creates.

math

Math Functions and flows functions for maths operations on data