Crate flowstdlib

flowstlib is a library of flows and functions that can be used from flows.

The flow and function definition are used at compile time when compile flows that reference it.

At run-time library entries can be of two types, indicated by their ImplementationLocator

Native - a native binary (containing) all functions is built and linked by a flow runner program (e.g. flowr) so that any function referenced by a flow is executed natively at native speeds. flowr offers the user control using this via the -n, --native command line option.

Wasm - the functions in the library are compiled to WASM files and located within the library at runtime by the flow runner. If either the library if not linked natively, or the -n, --native command line option is not used, when the function is referenced by a flow being run, it is loaded and executed in a WASM runtime.

Modules

control

functions for controlling data flow

data

functions for generic operations on data

errors

provides Error that other modules in this crate will use errors::*; to get access to everything error_chain creates.

fmt

functions for formatting data

manifest

Use manifest::get_manifest to get the natively/statically linked LibraryManifest for this library

math

functions for maths operations on data