Crate flowstdlib

Modules

control
Control functions A Module with functions for controlling data flow

data
Data functions

fmt
Formatting functions

img
Imaging functions

math
Maths functions

Structs

Error
The Error type.

Enums

ErrorKind
The kind of an error.

Traits

ResultExt
Additional methods for Result, for easy interaction with this crate.

Functions

get_manifest
Return the LibraryManifest for the functions in the library

Type Definitions

Result
Convenient wrapper around std::Result.