Module flowstdlib::data

Data functions

Control (//flowstdlib/data)

Some generic Processes that act on data.

Modules

accumulate
A module that implements accumulating an array from a number of elements

append
A module that implements String concatenation

buffer
A module that implements a data buffer

count
A module that counts data passed thru it

duplicate
A module that duplicates objects into an array of them

duplicate_rows
A module that duplicates the rows in an array

enumerate
A module that enumerates entries of an Array

info
A module with a function to get info about a Value

multiply_row
A module that does matrix row multiplication

ordered_split
A module that splits a String into an Array of Strings

remove
A module that removes elements from an array

sort
A module with a function to sort values into an ordered array of numbers

split
A module that splits a String into an Array of Strings

transpose
A module with a function for transposing a Matrix

zip
A module that zips two sets of data into a set of tuples of data